chore: [M3-10715] - Bump Vite from `7.1.x` to `7.2.x` and update Vitest from `v3` to `v4` (#13119)
* chore: update vite to latest and vitest to v4

* fix search tests and sdk build

* addess unhanded promises in `TriggerConditions.test.tsx`

* fix some circular imports in CloudPulse to try to fix unit test stability

* lint

* clean up

* clean up

* fix lockfile

* fix unawaited fetches by mocking flags differently to hopefully fix test flake

* fix unawaited fetches by mocking flags differently to hopefully fix test flake

* mock flags with our built-in helper

* revert some possibly unnessesary changes

* lint

* add some changesets
