+++ package.json
@@ -8,8 +8,8 @@
     "concurrently": "9.1.0",
     "husky": "^9.1.6",
     "typescript": "^5.7.3",
-    "vitest": "^3.2.4",
-    "@vitest/ui": "^3.2.4",
+    "vitest": "^4.0.10",
+    "@vitest/ui": "^4.0.10",
     "lint-staged": "^15.4.3",
     "eslint": "^9.24.0",
     "eslint-config-prettier": "^10.1.1",
