+++ packages/api-v4/.changeset/pr-13119-tech-stories-1764007445902.md
@@ -1,0 +1,5 @@
+---
+"@linode/api-v4": Tech Stories
+---
+
+Add `@types/node` as a devDependency ([#13119](https://github.com/linode/manager/pull/13119))
