+++ packages/manager/.changeset/pr-13119-tech-stories-1764007617066.md
@@ -1,0 +1,5 @@
+---
+"@linode/manager": Tech Stories
+---
+
+Update vitest from `v3` to `v4` ([#13119](https://github.com/linode/manager/pull/13119))
