+++ packages/manager/package.json
@@ -180,7 +180,7 @@
     "pdfreader": "^3.0.7",
     "redux-mock-store": "^1.5.3",
     "storybook": "^9.0.12",
-    "vite": "^7.1.11",
+    "vite": "^7.2.2",
     "vite-plugin-svgr": "^4.5.0"
   },
   "browserslist": [
