+++ packages/manager/src/features/CloudPulse/Alerts/CreateAlert/Criteria/DimensionFilterValue/utils.test.ts
@@ -1,6 +1,5 @@
 import { linodeFactory, nodeBalancerFactory } from '@linode/utilities';
 
-import { transformDimensionValue } from '../../../Utils/utils';
 import {
   getFilteredFirewallParentEntities,
   getFirewallLinodes,
@@ -11,6 +10,7 @@ import {
   handleValueChange,
   resolveSelectedValues,
   scopeBasedFilteredResources,
+  transformDimensionValue,
 } from './utils';
 
 import type { Linode } from '@linode/api-v4';
