+++ packages/manager/src/features/CloudPulse/Utils/FilterBuilder.ts
@@ -6,9 +6,8 @@ import {
   RESOURCES,
   TAGS,
 } from './constants';
-import { FILTER_CONFIG } from './FilterConfig';
+import { FILTER_CONFIG, getAssociatedEntityType } from './FilterConfig';
 import { CloudPulseAvailableViews, CloudPulseSelectTypes } from './models';
-import { getAssociatedEntityType } from './utils';
 
 import type {
   CloudPulseMetricsFilter,
@@ -674,72 +673,6 @@ const getDependentFiltersByFilterKey = (
         ? RESOURCES
         : configuration.filterKey
     );
-};
-
-/**
- * @param obj1 The first object to be compared
- * @param obj2 The second object to be compared
- * @returns True if, both are equal else false
- */
-export const deepEqual = <T>(obj1: T, obj2: T): boolean => {
-  if (obj1 === obj2) {
-    return true; // Identical references or values
-  }
-
-  // If either is null or undefined, or they are not of object type, return false
-  if (
-    obj1 === null ||
-    obj2 === null ||
-    typeof obj1 !== 'object' ||
-    typeof obj2 !== 'object'
-  ) {
-    return false;
-  }
-
-  // Handle array comparison separately
-  if (Array.isArray(obj1) && Array.isArray(obj2)) {
-    return compareArrays(obj1, obj2);
-  }
-
-  // Ensure both objects have the same number of keys
-  const keys1 = Object.keys(obj1);
-  const keys2 = Object.keys(obj2);
-
-  if (keys1.length !== keys2.length) {
-    return false;
-  }
-
-  // Recursively check each key
-  for (const key of keys1) {
-    if (!(key in obj2)) {
-      return false;
-    }
-    // Recursive deep equal check
-    if (!deepEqual((obj1 as any)[key], (obj2 as any)[key])) {
-      return false;
-    }
-  }
-
-  return true;
-};
-
-/**
- * @param arr1 Array for comparison
- * @param arr2 Array for comparison
- * @returns True if, both the arrays are equal, else false
- */
-export const compareArrays = <T>(arr1: T[], arr2: T[]): boolean => {
-  if (arr1.length !== arr2.length) {
-    return false;
-  }
-
-  for (let i = 0; i < arr1.length; i++) {
-    if (!deepEqual(arr1[i], arr2[i])) {
-      return false;
-    }
-  }
-
-  return true;
 };
 
 /**
