+++ packages/manager/src/utilities/analytics/utils.test.ts
@@ -120,13 +120,13 @@ describe('waitForAdobeAnalyticsToBeLoaded', () => {
 
   it(
     'should reject if adobe is not defined after 5 seconds',
+    { timeout: 7000 },
     () => {
       vi.stubGlobal('_satellite', undefined);
       expect(waitForAdobeAnalyticsToBeLoaded()).rejects.toThrow(
         'Adobe Analytics did not load after 5 seconds'
       );
-    },
-    { timeout: 7000 }
+    }
   );
 });
