+++ packages/manager/vite.config.ts
@@ -28,6 +28,7 @@ export default defineConfig({
     port: 3000,
   },
   test: {
+    include: ['**/*.test.{js,jsx,ts,tsx}'],
     coverage: {
       exclude: [
         'src/**/*.constants.{js,jsx,ts,tsx}',
@@ -44,7 +45,6 @@ export default defineConfig({
     },
     environment: 'jsdom',
     globals: true,
-    pool: 'forks',
     setupFiles: './src/testSetup.ts',
   },
 });
