+++ packages/utilities/.changeset/pr-13119-tech-stories-1764007518381.md
@@ -1,0 +1,5 @@
+---
+"@linode/utilities": Tech Stories
+---
+
+Update `scrollErrorIntoViewV2.test.tsx‎` to not mock  MutationObserver` ([#13119](https://github.com/linode/manager/pull/13119))
