+++ vitest.config.ts
@@ -1,0 +1,14 @@
+import { defineConfig } from "vitest/config";
+
+export default defineConfig({
+  test: {
+    projects: [
+      "packages/api-v4",
+      "packages/manager",
+      "packages/search",
+      "packages/shared",
+      "packages/ui",
+      "packages/utilities",
+    ],
+  },
+});
