--- vitest.workspace.ts
@@ -1,8 +1,0 @@
-export default [
-  "packages/api-v4",
-  "packages/manager",
-  "packages/search",
-  "packages/shared",
-  "packages/ui",
-  "packages/utilities",
-];
